Render otimizado, Renderiza somente o que vai ser mostrado e não tudo.

+++ src/br/udesc/ceavi/pin/modulo1/control/funtion/FuntionMoveTela.java
@@ -43,14 +43,14 @@ public class FuntionMoveTela implements IMoveTelaFuntion {
 
                 /// (-tela.size / (2 * Zoom) <-> tela.size / (2 * Zoom))   
                 //Não Precisão Na Porção da Tela A ser motrado pelo sistema Como ZOOM Aplicado
-                int limiteWidth = HelpLocator.getNetworkWidth() / (2 * HelpLocator.getZOOM());
+                int limiteWidth = (int) (HelpLocator.getNetworkWidth() / (2 * HelpLocator.getZOOM()));
                 boolean test1Width = (HelpLocator.getGuideX() + xMove * 0.2) > -limiteWidth;
                 boolean test2Width = (HelpLocator.getGuideX() + xMove * 0.2) < limiteWidth;
                 if (test1Width && test2Width) {
                     HelpLocator.setGuiaX((int) (HelpLocator.getGuideX() + xMove * 0.2));
                 }
 
-                int limiteHeight = HelpLocator.getNetworkHeight() / (2 * HelpLocator.getZOOM());
+                int limiteHeight = (int) (HelpLocator.getNetworkHeight() / (2 * HelpLocator.getZOOM()));
                 boolean test1Height = (HelpLocator.getGuideY() + yMove * 0.2) > -limiteHeight;
                 boolean test2Height = (HelpLocator.getGuideY() + yMove * 0.2) < limiteHeight;
                 if (test1Height && test2Height) {
